Knopki Post i Comment, pochti v polnom sostave, krome tech gde est` publishDate.

+++ src/main/java/ru/skillbox/zerone/backend/model/dto/response/CommentDTO.java
@@ -8,12 +8,12 @@ import java.util.List;
 @Data
 public class CommentDTO {
   @JsonProperty("parent_id")
-  private Integer parentId;
+  private Long parentId;
   @JsonProperty("comment_text")
   private String commentText;
-  private int id;
+  private long id;
   @JsonProperty("post_id")
-  private int postId;
+  private long postId;
   private LocalDateTime time;
   private UserDTO author;
   @JsonProperty("is_blocked")
